refactor(testing): merge the verify-proof fixtures and deduplicate aggregation
The two proof-verification fixture files duplicated the alternate-head
sentinel, three tamper classes, and the false-positive swap guard.
The single-message vector is the degenerate one-component case of the
multi-message vector, so the tamper vocabulary is now shared: one set
of five tamper classes where component_index defaults to 0, with the
single-message fixture accepting the three that apply to it.

The identical leaf-aggregation idiom (sign each validator, zip with
public keys, aggregate into a single-message proof) appeared four
times: twice across the verify fixtures and twice in the key manager.
The existing sign-and-aggregate method already implemented it, so it
gains an optional precomputed-signature lookup and all four sites
collapse onto it. The only remaining raw aggregation call is the
recursive two-level fold, which is genuinely unique.

Changes:

- one verify_proofs module replaces the two fixture files; the
  sentinel, tamper classes, component-index guard, alternate-head
  rebind, and participant-key swap guard exist once
- tamper renames: the Component-prefixed variants disappear
  (RebindComponentToAlternateHeadRoot -> RebindToAlternateHeadRoot,
  IncrementComponentSlot -> IncrementEmittedSlot,
  SwapComponentParticipantPublicKey -> SwapParticipantPublicKey,
  SwapComponentMessageBindings -> SwapMessageBindings,
  DropComponentMessageBinding -> DropMessageBinding); the single
  spec's bare index field becomes participant_index
- the key manager's attestation-proof builder delegates to
  sign-and-aggregate instead of repeating the aggregation idiom

Generated vectors are unchanged: the fixtures emit the same fields
and the aggregation inputs are byte-identical.
