+++ packages/testing/src/consensus_testing/__init__.py
@@ -7,21 +7,18 @@ from consensus_testing.genesis import build_anchor, generate_pre_state
 from consensus_testing.test_fixtures import (
     ApiEndpointTest,
     BaseConsensusFixture,
-    DropComponentMessageBinding,
+    DropMessageBinding,
     ForkChoiceTest,
     GossipsubHandlerTest,
-    IncrementComponentSlot,
     IncrementEmittedSlot,
     JustifiabilityTest,
     NetworkingCodecTest,
     PoseidonPermutationTest,
-    RebindComponentToAlternateHeadRoot,
     RebindToAlternateHeadRoot,
     SlotClockTest,
     SSZTest,
     StateTransitionTest,
-    SwapComponentMessageBindings,
-    SwapComponentParticipantPublicKey,
+    SwapMessageBindings,
     SwapParticipantPublicKey,
     SyncTest,
     VerifyMultiMessageProofsTest,
@@ -72,15 +69,12 @@ __all__ = [
     "StateTransitionTest",
     "ForkChoiceTest",
     "VerifySingleMessageProofsTest",
+    "VerifyMultiMessageProofsTest",
     "RebindToAlternateHeadRoot",
     "IncrementEmittedSlot",
     "SwapParticipantPublicKey",
-    "VerifyMultiMessageProofsTest",
-    "RebindComponentToAlternateHeadRoot",
-    "IncrementComponentSlot",
-    "SwapComponentParticipantPublicKey",
-    "SwapComponentMessageBindings",
-    "DropComponentMessageBinding",
+    "SwapMessageBindings",
+    "DropMessageBinding",
     "VerifySignaturesTest",
     "SSZTest",
     "NetworkingCodecTest",
